scripts/sitemap/requirements.txt
# SPDX-FileCopyrightText: 2024 Ondsel <development@ondsel.com>
#
# SPDX-License-Identifier: AGPL-3.0-or-later

pymongo
boto3
python-dotenv